[0.3.11] (破坏性更新) 令SQLUpdateAction返回的值为 Long 以适配自增主键大小范围。

+++ example/demo/src/test/java/cc/carm/lib/easysql/tests/SQLUpdateBatchTests.java
@@ -25,7 +25,7 @@ public class SQLUpdateBatchTests extends TestHandler {
     @Override
     public void onTest(SQLManager sqlManager) throws SQLException {
 
-        List<Integer> updates = sqlManager.createInsertBatch("test_user_table")
+        List<Long> updates = sqlManager.createInsertBatch("test_user_table")
                 .setColumnNames("uuid", "username", "age")
                 .setAllParams(generateParams())
                 .execute();
